added TickTimer, changed_by, fixed transform sync bug
TicksTimer is a resources which stores a btreemap of Tick -> elapsed
Duration, frame count
It filters itself to reduce resolution for older ticks. Grows as log(elapsed) and has around
800 entries after a few minutes.

Added changed_by with link to open in vscode.

In the process of the above, the bug where rigidbody started moving
erratically after being selected in editor seems to be gone. I can't
replicate it.